Include row and colnames
First submission (at SWC instructor training Melbourne) - new to git, md and Rmd

Row and Col names are incredibly useful to researchers and make it easier to understand subsets later in course. I know this will change a substantial number of outputs. Is there a way to generate the output again?

Any help or feedback on whether this is feasible or beneficial would be appreciated